feat: add ActionButtons, CountdownEdit, MediaButtons, ToolButtons components and update timeTabTypes
