+++ src/renderer/components/states/hover/pages/time/types/timeTabTypes.ts
@@ -31,3 +31,17 @@ export interface TimeTabProps {
   /** 农历日期字符串 */
   lunarStr: string;
 }
+
+/** ActionButtons 组件 Props */
+export interface ActionButtonsProps {
+  /** 隐藏图标路径 */
+  hideIcon?: string;
+  /** 退出图标路径 */
+  powerOffIcon?: string;
+}
+
+/** 截图模式 */
+export type HoverScreenshotMode = 'region' | 'display';
+
+/** 计时器状态 */
+export type TimerState = 'idle' | 'running' | 'paused';
